fix: related item identifier endpoint
